requirements.txt
fastapi
requests
uvicorn
python-dotenv
psycopg2-binary